Add Authorization Context
This context to manage and keep track of the user authentication. Notice that it contains isAuthenticated, userName, and authToken state variables. It also provides an authanticate() and register() functions that use the API to authenticate and register users. The localStorage.setItem("token", data); statement is used to store the JWT token in the browsers local storage.